Generate more from the manifest, delete what only re-proved the type system (#1066)

+++ packages/schemas/src/problems.ts
@@ -1118,9 +1118,9 @@ const byTypeFields = Object.fromEntries(
 
 export const problemsCountSchema = z.object({
   /** Defect rows only — what the navbar badge and homepage banner show. */
-  total: z.number(),
+  total: z.number().int(),
   /** Coverage-backlog rows, reported separately so they never inflate `total`. */
-  coverageTotal: z.number(),
+  coverageTotal: z.number().int(),
   byType: z.object(byTypeFields),
 });
 
@@ -1192,9 +1192,9 @@ export const countProblems = (all: AllProblems): ProblemsCount => {
 };
 
 // Assemble the cost-grouped detector results into the combined AllProblems
-// shape (with derived total). Shared by the service-layer findAllProblems
-// aggregator and the MCP list_problems tool so the merge + total live in one
-// place. (The Problems page merges client-side in useProblemsData, which is
+// shape (with derived total). Shared so the merge + total live in one place —
+// the MCP list_problems tool calls this directly over its own per-group
+// scans. (The Problems page merges client-side in useProblemsData, which is
 // loading-aware and defaults not-yet-loaded groups to empty.)
 export const assembleAllProblems = (groups: {
   fast: ProblemsFast;
